dev: revamp app sidebar
