+++ web/app/[workspaceSlug]/(projects)/sidebar.tsx
@@ -1,18 +1,21 @@
 import { FC, useRef } from "react";
 import { observer } from "mobx-react";
 // components
-import { ProjectSidebarList } from "@/components/project";
 import {
-  WorkspaceHelpSection,
-  WorkspaceSidebarDropdown,
-  WorkspaceSidebarMenu,
-  WorkspaceSidebarQuickAction,
+  SidebarDropdown,
+  SidebarHelpSection,
+  SidebarProjectsList,
+  SidebarQuickActions,
+  SidebarUserMenu,
+  SidebarWorkspaceMenu,
 } from "@/components/workspace";
+// helpers
+import { cn } from "@/helpers/common.helper";
 // hooks
 import { useAppTheme } from "@/hooks/store";
 import useOutsideClickDetector from "@/hooks/use-outside-click-detector";
 
-export interface IAppSidebar { }
+export interface IAppSidebar {}
 
 export const AppSidebar: FC<IAppSidebar> = observer(() => {
   // store hooks
@@ -30,19 +33,20 @@ export const AppSidebar: FC<IAppSidebar> = observer(() => {
 
   return (
     <div
-      className={`fixed inset-y-0 z-20 flex h-full flex-shrink-0 flex-grow-0 flex-col border-r border-custom-sidebar-border-200 bg-custom-sidebar-background-100
-        duration-300 md:relative
-        ${sidebarCollapsed ? "-ml-[250px]" : ""}
-        sm:${sidebarCollapsed ? "-ml-[250px]" : ""}
-        md:ml-0 ${sidebarCollapsed ? "w-[70px]" : "w-[250px]"}
-      `}
+      className={cn(
+        "fixed inset-y-0 z-20 flex h-full flex-shrink-0 flex-grow-0 flex-col border-r border-custom-sidebar-border-200 bg-custom-sidebar-background-100 duration-300 w-[250px] md:relative md:ml-0",
+        {
+          "w-[70px] -ml-[250px]": sidebarCollapsed,
+        }
+      )}
     >
-      <div ref={ref} className="flex h-full w-full flex-1 flex-col">
-        <WorkspaceSidebarDropdown />
-        <WorkspaceSidebarQuickAction />
-        <WorkspaceSidebarMenu />
-        <ProjectSidebarList />
-        <WorkspaceHelpSection />
+      <div ref={ref} className="size-full flex flex-col flex-1 gap-y-4 p-4 pb-0">
+        <SidebarDropdown />
+        <SidebarQuickActions />
+        <SidebarUserMenu />
+        <SidebarWorkspaceMenu />
+        <SidebarProjectsList />
+        <SidebarHelpSection />
       </div>
     </div>
   );
