+++ web/core/components/project/index.ts
@@ -13,8 +13,6 @@ export * from "./form";
 export * from "./join-project-modal";
 export * from "./leave-project-modal";
 export * from "./member-select";
-export * from "./sidebar-list-item";
-export * from "./sidebar-list";
 export * from "./integration-card";
 export * from "./member-list";
 export * from "./member-list-item";
