+++ web/core/components/workspace/index.ts
@@ -1,12 +1,9 @@
 export * from "./settings";
+export * from "./sidebar";
 export * from "./views";
 export * from "./confirm-workspace-member-remove";
 export * from "./create-workspace-form";
 export * from "./delete-workspace-modal";
-export * from "./help-section";
 export * from "./logo";
 export * from "./send-workspace-invitation-modal";
-export * from "./sidebar-dropdown";
-export * from "./sidebar-menu";
-export * from "./sidebar-quick-action";
 export * from "./workspace-active-cycles-upgrade";
