+++ web/core/components/workspace/sidebar/index.ts
@@ -1,0 +1,7 @@
+export * from "./dropdown";
+export * from "./help-section";
+export * from "./projects-list-item";
+export * from "./projects-list";
+export * from "./quick-actions";
+export * from "./user-menu";
+export * from "./workspace-menu";
