+++ web/core/components/workspace/sidebar/projects-list.tsx
@@ -1,0 +1,298 @@
+"use client";
+
+import { useState, FC, useRef, useEffect } from "react";
+import { combine } from "@atlaskit/pragmatic-drag-and-drop/combine";
+import { autoScrollForElements } from "@atlaskit/pragmatic-drag-and-drop-auto-scroll/element";
+import { observer } from "mobx-react";
+import { useParams } from "next/navigation";
+import { ChevronUp, Plus } from "lucide-react";
+import { Disclosure, Transition } from "@headlessui/react";
+// types
+import { IProject } from "@plane/types";
+// ui
+import { TOAST_TYPE, setToast } from "@plane/ui";
+// components
+import { CreateProjectModal } from "@/components/project";
+import { SidebarProjectsListItem } from "@/components/workspace";
+// constants
+import { EUserWorkspaceRoles } from "@/constants/workspace";
+// helpers
+import { cn } from "@/helpers/common.helper";
+import { orderJoinedProjects } from "@/helpers/project.helper";
+import { copyUrlToClipboard } from "@/helpers/string.helper";
+// hooks
+import { useAppTheme, useCommandPalette, useEventTracker, useProject, useUser } from "@/hooks/store";
+
+export const SidebarProjectsList: FC = observer(() => {
+  // get local storage data for isFavoriteProjectsListOpen and isAllProjectsListOpen
+  const isFavProjectsListOpenInLocalStorage = localStorage.getItem("isFavoriteProjectsListOpen");
+  const isAllProjectsListOpenInLocalStorage = localStorage.getItem("isAllProjectsListOpen");
+  // states
+  const [isFavoriteProjectsListOpen, setIsFavoriteProjectsListOpen] = useState(
+    isFavProjectsListOpenInLocalStorage === "true"
+  );
+  const [isAllProjectsListOpen, setIsAllProjectsListOpen] = useState(isAllProjectsListOpenInLocalStorage === "true");
+  const [isFavoriteProjectCreate, setIsFavoriteProjectCreate] = useState(false);
+  const [isProjectModalOpen, setIsProjectModalOpen] = useState(false);
+  const [isScrolled, setIsScrolled] = useState(false); // scroll animation state
+  // refs
+  const containerRef = useRef<HTMLDivElement | null>(null);
+  // store hooks
+  const { toggleCreateProjectModal } = useCommandPalette();
+  const { sidebarCollapsed } = useAppTheme();
+  const { setTrackElement } = useEventTracker();
+  const {
+    membership: { currentWorkspaceRole },
+  } = useUser();
+  const {
+    getProjectById,
+    joinedProjectIds: joinedProjects,
+    favoriteProjectIds: favoriteProjects,
+    updateProjectView,
+  } = useProject();
+  // router params
+  const { workspaceSlug } = useParams();
+  // auth
+  const isAuthorizedUser = !!currentWorkspaceRole && currentWorkspaceRole >= EUserWorkspaceRoles.MEMBER;
+
+  const handleCopyText = (projectId: string) => {
+    copyUrlToClipboard(`${workspaceSlug}/projects/${projectId}/issues`).then(() => {
+      setToast({
+        type: TOAST_TYPE.SUCCESS,
+        title: "Link Copied!",
+        message: "Project link copied to clipboard.",
+      });
+    });
+  };
+
+  const handleOnProjectDrop = (
+    sourceId: string | undefined,
+    destinationId: string | undefined,
+    shouldDropAtEnd: boolean
+  ) => {
+    if (!sourceId || !destinationId || !workspaceSlug) return;
+    if (sourceId === destinationId) return;
+
+    const joinedProjectsList: IProject[] = [];
+    joinedProjects.map((projectId) => {
+      const projectDetails = getProjectById(projectId);
+      if (projectDetails) joinedProjectsList.push(projectDetails);
+    });
+
+    const sourceIndex = joinedProjects.indexOf(sourceId);
+    const destinationIndex = shouldDropAtEnd ? joinedProjects.length : joinedProjects.indexOf(destinationId);
+
+    if (joinedProjectsList.length <= 0) return;
+
+    const updatedSortOrder = orderJoinedProjects(sourceIndex, destinationIndex, sourceId, joinedProjectsList);
+    if (updatedSortOrder != undefined)
+      updateProjectView(workspaceSlug.toString(), sourceId, { sort_order: updatedSortOrder }).catch(() => {
+        setToast({
+          type: TOAST_TYPE.ERROR,
+          title: "Error!",
+          message: "Something went wrong. Please try again.",
+        });
+      });
+  };
+
+  const isCollapsed = sidebarCollapsed || false;
+
+  /**
+   * Implementing scroll animation styles based on the scroll length of the container
+   */
+  useEffect(() => {
+    const handleScroll = () => {
+      if (containerRef.current) {
+        const scrollTop = containerRef.current.scrollTop;
+        setIsScrolled(scrollTop > 0);
+      }
+    };
+    const currentContainerRef = containerRef.current;
+    if (currentContainerRef) {
+      currentContainerRef.addEventListener("scroll", handleScroll);
+    }
+    return () => {
+      if (currentContainerRef) {
+        currentContainerRef.removeEventListener("scroll", handleScroll);
+      }
+    };
+  }, [containerRef]);
+
+  useEffect(() => {
+    const element = containerRef.current;
+
+    if (!element) return;
+
+    return combine(
+      autoScrollForElements({
+        element,
+        canScroll: ({ source }) => source?.data?.dragInstanceId === "PROJECTS",
+        getAllowedAxis: () => "vertical",
+      })
+    );
+  }, [containerRef]);
+
+  const toggleListDisclosure = (isOpen: boolean, type: "all" | "favorite") => {
+    if (type === "all") {
+      setIsAllProjectsListOpen(isOpen);
+      localStorage.setItem("isAllProjectsListOpen", isOpen.toString());
+    } else {
+      setIsFavoriteProjectsListOpen(isOpen);
+      localStorage.setItem("isFavoriteProjectsListOpen", isOpen.toString());
+    }
+  };
+
+  return (
+    <>
+      {workspaceSlug && (
+        <CreateProjectModal
+          isOpen={isProjectModalOpen}
+          onClose={() => {
+            setIsProjectModalOpen(false);
+          }}
+          setToFavorite={isFavoriteProjectCreate}
+          workspaceSlug={workspaceSlug.toString()}
+        />
+      )}
+      <div
+        ref={containerRef}
+        className={cn("vertical-scrollbar h-full space-y-2 !overflow-y-scroll scrollbar-sm -mr-3", {
+          "border-t border-custom-sidebar-border-300": isScrolled,
+        })}
+      >
+        {favoriteProjects && favoriteProjects.length > 0 && (
+          <Disclosure as="div" className="flex flex-col" defaultOpen={isFavoriteProjectCreate}>
+            <>
+              {!isCollapsed && (
+                <div className="group w-full flex items-center justify-between px-2 py-0.5 rounded text-custom-sidebar-text-400 hover:bg-custom-sidebar-background-90">
+                  <Disclosure.Button
+                    as="button"
+                    type="button"
+                    className="group w-full flex items-center gap-1 whitespace-nowrap text-left text-sm font-medium text-custom-sidebar-text-400"
+                    onClick={() => toggleListDisclosure(!isFavoriteProjectsListOpen, "favorite")}
+                  >
+                    Favorites
+                    <ChevronUp
+                      className={cn("flex-shrink-0 size-3.5 transition-all", {
+                        "rotate-180": isFavoriteProjectsListOpen,
+                      })}
+                    />
+                  </Disclosure.Button>
+                  {isAuthorizedUser && (
+                    <button
+                      className="opacity-0 group-hover:opacity-100 p-0.5 rounded hover:bg-custom-sidebar-background-80"
+                      onClick={() => {
+                        setTrackElement("APP_SIDEBAR_FAVORITES_BLOCK");
+                        setIsFavoriteProjectCreate(true);
+                        setIsProjectModalOpen(true);
+                      }}
+                    >
+                      <Plus className="size-3" />
+                    </button>
+                  )}
+                </div>
+              )}
+              <Transition
+                show={isFavoriteProjectsListOpen}
+                enter="transition duration-100 ease-out"
+                enterFrom="transform scale-95 opacity-0"
+                enterTo="transform scale-100 opacity-100"
+                leave="transition duration-75 ease-out"
+                leaveFrom="transform scale-100 opacity-100"
+                leaveTo="transform scale-95 opacity-0"
+              >
+                {isFavoriteProjectsListOpen && (
+                  <Disclosure.Panel as="div" static>
+                    {favoriteProjects.map((projectId, index) => (
+                      <SidebarProjectsListItem
+                        key={projectId}
+                        projectId={projectId}
+                        handleCopyText={() => handleCopyText(projectId)}
+                        projectListType="FAVORITES"
+                        disableDrag
+                        disableDrop
+                        isLastChild={index === favoriteProjects.length - 1}
+                      />
+                    ))}
+                  </Disclosure.Panel>
+                )}
+              </Transition>
+            </>
+          </Disclosure>
+        )}
+        {joinedProjects && joinedProjects.length > 0 && (
+          <Disclosure as="div" defaultOpen={isAllProjectsListOpen}>
+            <>
+              {!isCollapsed && (
+                <div className="group w-full flex items-center justify-between px-2 py-0.5 rounded text-custom-sidebar-text-400 hover:bg-custom-sidebar-background-90">
+                  <Disclosure.Button
+                    as="button"
+                    type="button"
+                    className="group w-full flex items-center gap-1 whitespace-nowrap text-left text-sm font-medium text-custom-sidebar-text-400"
+                    onClick={() => toggleListDisclosure(!isAllProjectsListOpen, "all")}
+                  >
+                    My projects
+                    <ChevronUp
+                      className={cn("flex-shrink-0 size-3.5 transition-all", {
+                        "rotate-180": isAllProjectsListOpen,
+                      })}
+                    />
+                  </Disclosure.Button>
+                  {isAuthorizedUser && (
+                    <button
+                      className="opacity-0 group-hover:opacity-100 p-0.5 rounded hover:bg-custom-sidebar-background-80"
+                      onClick={() => {
+                        setTrackElement("Sidebar");
+                        setIsFavoriteProjectCreate(false);
+                        setIsProjectModalOpen(true);
+                      }}
+                    >
+                      <Plus className="size-3" />
+                    </button>
+                  )}
+                </div>
+              )}
+              <Transition
+                show={isAllProjectsListOpen}
+                enter="transition duration-100 ease-out"
+                enterFrom="transform scale-95 opacity-0"
+                enterTo="transform scale-100 opacity-100"
+                leave="transition duration-75 ease-out"
+                leaveFrom="transform scale-100 opacity-100"
+                leaveTo="transform scale-95 opacity-0"
+              >
+                {isAllProjectsListOpen && (
+                  <Disclosure.Panel as="div" className="mt-3" static>
+                    {joinedProjects.map((projectId, index) => (
+                      <SidebarProjectsListItem
+                        key={projectId}
+                        projectId={projectId}
+                        projectListType="JOINED"
+                        handleCopyText={() => handleCopyText(projectId)}
+                        isLastChild={index === joinedProjects.length - 1}
+                        handleOnProjectDrop={handleOnProjectDrop}
+                      />
+                    ))}
+                  </Disclosure.Panel>
+                )}
+              </Transition>
+            </>
+          </Disclosure>
+        )}
+        {isAuthorizedUser && joinedProjects && joinedProjects.length === 0 && (
+          <button
+            type="button"
+            className="flex w-full items-center gap-2 px-3 text-sm text-custom-sidebar-text-200"
+            onClick={() => {
+              setTrackElement("Sidebar");
+              toggleCreateProjectModal(true);
+            }}
+          >
+            <Plus className="h-5 w-5" />
+            {!isCollapsed && "Add Project"}
+          </button>
+        )}
+      </div>
+    </>
+  );
+});
