+++ web/core/components/workspace/sidebar/user-menu.tsx
@@ -4,22 +4,21 @@ import React from "react";
 import { observer } from "mobx-react";
 import Link from "next/link";
 import { useParams, usePathname } from "next/navigation";
-import { Crown } from "lucide-react";
 // ui
 import { Tooltip } from "@plane/ui";
 // components
 import { NotificationPopover } from "@/components/notifications";
 // constants
-import { SIDEBAR_MENU_ITEMS } from "@/constants/dashboard";
+import { SIDEBAR_USER_MENU_ITEMS } from "@/constants/dashboard";
 import { SIDEBAR_CLICKED } from "@/constants/event-tracker";
 import { EUserWorkspaceRoles } from "@/constants/workspace";
-// helper
+// helpers
 import { cn } from "@/helpers/common.helper";
 // hooks
 import { useAppTheme, useEventTracker, useUser } from "@/hooks/store";
 import { usePlatformOS } from "@/hooks/use-platform-os";
 
-export const WorkspaceSidebarMenu = observer(() => {
+export const SidebarUserMenu = observer(() => {
   // store hooks
   const { toggleSidebar, sidebarCollapsed } = useAppTheme();
   const { captureEvent } = useEventTracker();
@@ -44,12 +43,12 @@ export const WorkspaceSidebarMenu = observer(() => {
   };
 
   return (
-    <div className="w-full cursor-pointer space-y-2 p-4">
-      {SIDEBAR_MENU_ITEMS.map(
+    <div className="w-full space-y-1">
+      {SIDEBAR_USER_MENU_ITEMS.map(
         (link) =>
           workspaceMemberInfo >= link.access && (
             <Link key={link.key} href={`/${workspaceSlug}${link.href}`} onClick={() => handleLinkClick(link.key)}>
-              <span className="my-1 block w-full">
+              <span className="block w-full">
                 <Tooltip
                   tooltipContent={link.label}
                   position="right"
@@ -58,23 +57,17 @@ export const WorkspaceSidebarMenu = observer(() => {
                   isMobile={isMobile}
                 >
                   <div
-                    className={`group flex w-full items-center gap-2.5 rounded-md px-3 py-2 text-sm font-medium outline-none ${
-                      link.highlight(pathname, `/${workspaceSlug}`)
-                        ? "bg-custom-primary-100/10 text-custom-primary-100"
-                        : "text-custom-sidebar-text-200 hover:bg-custom-sidebar-background-80 focus:bg-custom-sidebar-background-80"
-                    } ${sidebarCollapsed ? "justify-center" : ""}`}
-                  >
-                    {
-                      <link.Icon
-                        className={cn("h-4 w-4", {
-                          "rotate-180": link.key === "active-cycles",
-                        })}
-                      />
-                    }
-                    {!sidebarCollapsed && <p className="leading-5">{link.label}</p>}
-                    {!sidebarCollapsed && link.key === "active-cycles" && (
-                      <Crown className="h-3.5 w-3.5 text-amber-400" />
+                    className={cn(
+                      "group w-full flex items-center gap-1.5 rounded-md px-2 py-1.5 outline-none text-custom-sidebar-text-200 hover:bg-custom-sidebar-background-90 focus:bg-custom-sidebar-background-90",
+                      {
+                        "text-custom-primary-100 bg-custom-primary-100/10 hover:bg-custom-primary-100/10":
+                          link.highlight(pathname, `/${workspaceSlug}`),
+                        "justify-center": sidebarCollapsed,
+                      }
                     )}
+                  >
+                    {<link.Icon className="size-4" />}
+                    {!sidebarCollapsed && <p className="text-sm leading-5 font-medium">{link.label}</p>}
                   </div>
                 </Tooltip>
               </span>
